+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.2.1
 - Fixed a bug where the phone wasn't closing with Tab, only Esc after using a slider
 - Fixed compatibility bug where LogManager would crash the Melon init on ML 0.7.2
+- Improved robustness of matching heuristics by checking the assembly name as well
+- Improved section matching in Log Explorer by checking assembly names and allowing partial matches
 ## 1.2.0
 - Improved UI visually by implementing panel corners
 - Added theme presets and custom theming options